Refactoring of deployment script
