+++ deploy/.env.example
@@ -1,18 +1,24 @@
+DEPLOYMENT_ACCOUNT_ADDRESS=0xb8988b690910913c97a090c3a6f80fad8b3a4683
+DEPLOYMENT_ACCOUNT_PRIVATE_KEY=67..14
+
 HOME_RPC_URL=https://sokol.poa.network
-HOME_PROXY_OWNER=0xb8988b690910913c97a090c3a6f80fad8b3a4683
-HOME_PROXY_OWNER_PRIVATE_KEY=6...z
-HOME_DAILY_LIMIT=1000000000000000000
-HOME_MAX_AMOUNT_PER_TX=100000000000000000
-HOME_MIN_AMOUNT_PER_TX=10000000000000000
+HOME_OWNER_MULTISIG=
+HOME_UPGRADEABLE_ADMIN_VALIDATORS=
+HOME_UPGRADEABLE_ADMIN_BRIDGE=
+HOME_DAILY_LIMIT=
+HOME_MAX_AMOUNT_PER_TX=
+HOME_MIN_AMOUNT_PER_TX=
 
 FOREIGN_RPC_URL=https://kovan.infura.io/mew
-FOREIGN_PROXY_OWNER=0xb8988b690910913c97a090c3a6f80fad8b3a4683
-FOREIGN_PROXY_OWNER_PRIVATE_KEY=6...z
-FOREIGN_DAILY_LIMIT=1000000000000000000
-FOREIGN_MAX_AMOUNT_PER_TX=100000000000000000
-FOREIGN_MIN_AMOUNT_PER_TX=10000000000000000
+FOREIGN_OWNER_MULTISIG=
+FOREIGN_UPGRADEABLE_ADMIN_VALIDATORS=
+FOREIGN_UPGRADEABLE_ADMIN_BRIDGE=
+FOREIGN_DAILY_LIMIT=
+FOREIGN_MAX_AMOUNT_PER_TX=
+FOREIGN_MIN_AMOUNT_PER_TX=
 
-REQUIRED_NUMBER_OF_VALIDATORS=1
-VALIDATORS="0xE0f0a5a0D8bB8DeBB3727A6FAD6e955483b9C208"
+REQUIRED_NUMBER_OF_VALIDATORS=
+VALIDATORS=""
 GET_RECEIPT_INTERVAL_IN_MILLISECONDS=3000
-GAS_PRICE=1
+
+GAS_PRICE=
