+++ deploy/README.md
@@ -1,27 +1,41 @@
 # How to use
+0. compile source contracts:
+```
+poa-parity
+
 1. create .env file
 ```bash
-REQUIRED_NUMBER_OF_VALIDATORS=1
-VALIDATORS="0x7A6a585dB8cDFa88B9e8403c054ec2e912E9D32E"
-GET_RECEIPT_INTERVAL_IN_MILLISECONDS=3000
+DEPLOYMENT_ACCOUNT_ADDRESS=0xb8988b690910913c97a090c3a6f80fad8b3a4683
+DEPLOYMENT_ACCOUNT_PRIVATE_KEY=67..14
 
 HOME_RPC_URL=https://sokol.poa.network
-HOME_PROXY_OWNER=0x08660156928d768D26D3eeF642c11527FDca0891
-HOME_PROXY_OWNER_PRIVATE_KEY=INSERT_PRIVATE_KEY_HERE
+HOME_OWNER_MULTISIG=0xdd2BcC1e053aBB1DfA6c1F3D6C7842f57d61440F
+HOME_UPGRADEABLE_ADMIN_VALIDATORS=0xdd2BcC1e053aBB1DfA6c1F3D6C7842f57d61440F
+HOME_UPGRADEABLE_ADMIN_BRIDGE=0xdd2BcC1e053aBB1DfA6c1F3D6C7842f57d61440F
 HOME_DAILY_LIMIT=1000000000000000000
 HOME_MAX_AMOUNT_PER_TX=100000000000000000
 HOME_MIN_AMOUNT_PER_TX=10000000000000000
+
 FOREIGN_RPC_URL=https://kovan.infura.io/mew
-FOREIGN_PROXY_OWNER=0x08660156928d768D26D3eeF642c11527FDca0891
-FOREIGN_PROXY_OWNER_PRIVATE_KEY=INSERT_PRIVATE_KEY_HERE
+FOREIGN_OWNER_MULTISIG=0xb8988b690910913c97a090c3a6f80fad8b3a4683
+FOREIGN_UPGRADEABLE_ADMIN_VALIDATORS=0xb8988b690910913c97a090c3a6f80fad8b3a4683
+FOREIGN_UPGRADEABLE_ADMIN_BRIDGE=0xb8988b690910913c97a090c3a6f80fad8b3a4683
 FOREIGN_DAILY_LIMIT=1000000000000000000
 FOREIGN_MAX_AMOUNT_PER_TX=100000000000000000
 FOREIGN_MIN_AMOUNT_PER_TX=10000000000000000
-GAS_PRICE=1
-```
 
+REQUIRED_NUMBER_OF_VALIDATORS=1
+VALIDATORS="0xb506698581484572b6ccfbd6c976f0948775eace"
+GET_RECEIPT_INTERVAL_IN_MILLISECONDS=3000
+
+GAS_PRICE=1
+
+
+```
+2.  Run `node deploy.js`
+OR
 2. Compile sources: `npm run compile`
-3. run `./build/deploy-macos` or `node deploy.js`
+3. run `./build/deploy-macos`
